#3 create Clean Architecture style projects
