+++ web-api/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "web-api"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+application = {path = "../application"}
+infrastructure = {path = "../infrastructure"}
